scoring changes
Invaders now get points for advancing.  The farther they
have advanced, the more points they get.

The player is now worth 5000 points instead of infinite.
Possibly the player could be worth 0, since the invaders
continue to accumulate points for advancing after the
player is dead.

The player loses a small number of points for misses.

As I commit this, I wonder how it would be if it worked
such that invaders get fewer points up front, so it's
easier for the player to stay ahead, until the player is
dead, at which point it's interesting to watch and see if
the invaders can catch up and overtake the player.  That
would fit the artistic theme of this whole thing. hmm.
(possibly the invaders could even lose points for misses,
as well... )