make it start easier but get harder each level
Start with a larger refire delay renage, decrease it
by 25% each level.  This means it starts out easier,
but by the 3rd level it should be slightly harder than
the original game.

Doing this partly because it's much harder to play on
a phone with the little joystick and button.