index.html is now build/index.html and built by script
index.html is no longer to be checked in.

It's now in the build directory, and only built if

Possibly I'm going to abandon this whole "build" idea.  I'm noticing
that fonts are loaded from the css, so I'd have to parse the css as
well. and blah blah.  and embedding binary data has been terrible
so far.  So maybe embrace the web philosophy and have nothing be
modular.